chore(telemetry): use static configs
rename file
