+++ ddtrace/settings/symbol_db.py
@@ -2,7 +2,7 @@ import re
 
 from envier import En
 
-from ddtrace.settings._core import report_telemetry as _report_telemetry
+from ddtrace.settings._telemetry import report_telemetry as _report_telemetry
 
 
 class SymbolDatabaseConfig(En):
